+++ Aurora.Core/IdGenerator.cs
