+++ Aurora.Core/InternalVariables.cs
