+++ Aurora.Core/Logger.cs
