+++ Aurora.Core/Owl.cs
