+++ Aurora.Core/SourceLocation.cs
