+++ Aurora.Core/Writer.cs
