+++ Aurora.Evaluator/BuiltinObjects/BooleanObject.cs
