+++ Aurora.Evaluator/BuiltinObjects/BooleanOutputStyleObject.cs
