+++ Aurora.Evaluator/BuiltinObjects/FloatObject.cs
