+++ Aurora.Evaluator/BuiltinObjects/OptionalObject.cs
