+++ Aurora.Evaluator/BuiltinObjects/StringObject.cs
