+++ Aurora.Evaluator/BuiltinObjects/TerminalObject.cs
