+++ Aurora.Evaluator/BuiltinObjects/TypeObject.cs
