+++ Aurora.Evaluator/EvaluatorRework.cs
