+++ Aurora.Evaluator/InternalMethods/Terminal.cs
