+++ Aurora.Evaluator/InternalMethods/Type.cs
