+++ Aurora.Evaluator/Internals/Attribute.cs
