+++ Aurora.Evaluator/Internals/Type.cs
