+++ Aurora.Evaluator/MathFunctions.cs
