+++ Aurora.Evaluator/variables.cs
