+++ Aurora/options.cs
@@ -1,4 +1,5 @@
 using System.CommandLine;
+using Aurora.Core;
 
 namespace Aurora;
 
@@ -164,35 +165,35 @@ public static class CommandLineArguments
 // {
 //     [Value(0, MetaName = "file", Required = true, HelpText = "Input file to be processed")]
 //     public required string FilePath { get; set; }
-// 
+//
 //     [Option("version", HelpText = "Display version information")]
 //     public bool Version { get; set; }
-// 
+//
 //     [Option('v', "verbose", Default = false, HelpText = "Prints verbose log messages")]
 //     public bool Verbose { get; set; }
-// 
+//
 //     [Option('d', "debug", Default = false, HelpText = "Prints debug log messages")]
 //     public bool Debug { get; set; }
-// 
+//
 //     [Option('w', "warn", Default = false, HelpText = "Prints warning log messages")]
 //     public bool Warning { get; set; }
-// 
+//
 //     [Option("no-console", Default = false, HelpText = "Suppresses log messages to the terminal")]
 //     public bool NoConsole { get; set; }
-// 
+//
 //     [Option("logfile", HelpText = "The file to output the log messages to")]
 //     public string? LogFile { get; set; }
-// 
+//
 //     [Option('s', "strict", Default = false, HelpText = "Only displays the log messages specified from flags")]
 //     public bool Strict { get; set; }
-// 
+//
 //     [Option("inline-stack-trace", Default = false, HelpText = "Prints inline stack traces for system errors")]
 //     public bool InlineStackTrace { get; set; }
-// 
+//
 //     [Option("disable-easter-eggs", Default = false,
 //         HelpText = "Disables the easter eggs hidden within the Aurora interpreter")]
 //     public bool DisableEasterEggs { get; set; }
-// 
+//
 //     [Option("config-file", Default = null, HelpText = "The config file for the aurora interpreter")]
 //     public string? ConfigFile { get; set; }
 // }
